refactor: reorganize imports and correct doc error
